Renaming token module to syntax.
